requirements.txt
google-generativeai
sentence-transformers
faiss-cpu
numpy
python-dotenv
langchain-text-splitters
tiktoken 
discord